third_party/blink/web_tests/http/tests/devtools/elements/edit/edit-dom-actions-1-expected.txt

Tests that user can mutate DOM by means of elements panel.


Running: testSetUp

Running: testRemove
==== before ====
- <div id="testRemove">
      <div id="node-to-remove"></div>
  </div>
==== after ====
  <div id="testRemove"></div>

Running: testSetNodeName
==== before ====
- <div id="testSetNodeName">
      <div id="node-to-set-name"></div>
  </div>
==== after ====
- <div id="testSetNodeName">
      <span id="node-to-set-name" ></span>
  </div>

Running: testSetNodeNameInput
==== before ====
- <div id="testSetNodeNameInput">
      <div id="node-to-set-name-input"></div>
  </div>
==== after ====
- <div id="testSetNodeNameInput">
      <input id="node-to-set-name-input" >
  </div>

Running: testSetNodeValue
==== before ====
- <div id="testSetNodeValue">
      <div id="node-to-set-value">\n                Text\n              </div>
  </div>
==== after ====
- <div id="testSetNodeValue">
      <div id="node-to-set-value">  \n  Edited Text  \n  </div>
  </div>
